move main locations content to event maps
